Here is the code for Ulyses
To run the procedure on the first n letters (instead or running on all of them) you should modify the 'index> ...' part on the bottom part of the code. Putting the iteration you want to stop. This is the second code, so it is a broader version of the code used to scrape the 120k